Move api check routes out of app.py via blueprints
